% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beetleTreeInteractions.R
\docType{data}
\name{beetleTreeInteractions}
\alias{beetleTreeInteractions}
\title{Beetle tree interaction table.}
\format{
A data frame with 565 rows and 2 variables:
\describe{
  \item{beetle}{Species of bark beetles}
  \item{genus}{Genus of host tree of bark beetles}
}
}
\source{
\url{https://www.barkbeetles.info/}
}
\usage{
beetleTreeInteractions
}
\description{
A dataset of beetles and genues of host trees. Original information
is avaliable in \url{https://www.barkbeetles.info/}
}
\details{
We use this interaction tables to construct an incidence matrix.
This is a table representation of a bipartite graph
}
\keyword{datasets}
