% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_matrix.R
\name{bin_matrix}
\alias{bin_matrix}
\title{Generate a binary random matrix with a sucsess parameter}
\usage{
bin_matrix(m, n, succes)
}
\arguments{
\item{m}{The number of columns}

\item{n}{The number of rows}

\item{succes}{The succes rate in each row. Default is 0.5}
}
\description{
Generate a binary random matrix with a sucsess parameter
}
\examples{
bin_matrix (m = 3  , n =  4, succes =  0.3)
}

