% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bin_matrix.R
\name{get_bin_matrix}
\alias{get_bin_matrix}
\title{Generate a binary random matrix of ones and zeros with a
threshold parameter}
\usage{
get_bin_matrix(m, n, success)
}
\arguments{
\item{m}{The number of columns}

\item{n}{The number of rows}

\item{success}{The probability of success in each trial. Default is 0.5}
}
\value{
Matrix with ones and zeros
}
\description{
Generate a binary random matrix of ones and zeros with a
threshold parameter
}
\examples{
get_bin_matrix(m = 3  , n =  4, success = 0.3)
}
