% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incidence_matrix.R
\name{get_incidence_matrix}
\alias{get_incidence_matrix}
\title{Generate an incidence matrix from an interaction table}
\usage{
get_incidence_matrix(db, returnDataFrame = FALSE)
}
\arguments{
\item{db}{A data.frame with two columns.}

\item{returnDataFrame}{Logical. If TRUE returns a tibble. Default FALSE.}
}
\value{
An incidence matrix or tibble.
}
\description{
Deprecated. Use \code{\link{prepare_incidence_matrix}} instead.
}
\examples{
\dontrun{
get_incidence_matrix(beetleTreeInteractions)
}
}
