% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_reg_coefficients.R
\name{get_log_reg_coefficients}
\alias{get_log_reg_coefficients}
\title{Logistic regression coefficients}
\usage{
get_log_reg_coefficients(incidence, phydist)
}
\arguments{
\item{incidence}{Incidence binary matrix.}

\item{phydist}{Phylogenetic distance matrix.}
}
\value{
A data.frame of logistic regression coefficients.
}
\description{
Deprecated. Use \code{\link{fit_geotax_model}} instead.
}
\examples{
\dontrun{
incidence_matrix <- get_incidence_matrix(beetleTreeInteractions)
incidence_matrix <- incidence_matrix[, colnames(phy_dist)]
get_log_reg_coefficients(incidence_matrix, phy_dist)
}
}
