% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomical_tree.R
\name{get_taxonomical_tree}
\alias{get_taxonomical_tree}
\title{Create a taxonomic tree (cladogram) given a taxonomic table}
\usage{
get_taxonomical_tree(tax_table, power = 1)
}
\arguments{
\item{tax_table}{A data frame with taxonomic information.}

\item{power}{Numeric; power for branch length computation. Default is 1.}
}
\value{
A cladogram (phylo object).
}
\description{
Deprecated. Use \code{\link{prepare_taxonomic_tree}} instead.
}
\examples{
\dontrun{
get_taxonomical_tree(tax_table)
}
}
