% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_points.R
\docType{data}
\name{host_points}
\alias{host_points}
\title{Presence points of 50,000 tree genus.}
\format{
A data frame with 58764 rows and 2 variables:
\describe{
  \item{genus}{genus of trees.}
  \item{longitude}{longitude, in decimal degrees}
  \item{latitude}{latitude, in decimal degrees}
}
}
\source{
\url{https://www.barkbeetles.info/}
}
\usage{
host_points
}
\description{
A dataset containing the presence points of almost 59,000
presence points of tree genus. This genus are
potential host of bark beetles.
}
\keyword{datasets}
