% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_points_mexico.R
\docType{data}
\name{host_points_mexico}
\alias{host_points_mexico}
\title{Presence points of 3,200 bark beetle genus in Mexico.}
\format{
A data frame with 3254 rows and 3 variables:
\describe{
  \item{genus}{genus of the bark beetle.}
  \item{longitude}{longitude, in decimal degrees}
  \item{latitude}{latitude, in decimal degrees}
}
}
\source{
\url{https://www.barkbeetles.info/}
}
\usage{
host_points_mexico
}
\description{
A dataset containing the presence points of almost 3,254
presence points of host trees by genera of bark beetles.
}
\keyword{datasets}
