% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_points_sample.R
\docType{data}
\name{host_points_sample}
\alias{host_points_sample}
\title{Presence points of 1,000 bark beetle genus.}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
  \item{genus}{genus of the bark beetle.}
  \item{longitude}{longitude, in decimal degrees}
  \item{latitude}{latitude, in decimal degrees}
}
}
\source{
\url{https://www.barkbeetles.info/}
}
\usage{
host_points_sample
}
\description{
A dataset containing the presence points of almost 1,000
presence points of bark beetles by genera.
}
\keyword{datasets}
