% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_tree.R
\docType{data}
\name{host_tree}
\alias{host_tree}
\title{tree_host pylogenetic tree on geotax}
\format{An object of class \code{"phylo"}}
\usage{
data(host_tree)
}
\description{
Host phylogenetic tree with branch lengths in My.
It was calculated using phylomatic.
}
\examples{
data(host_tree)
dist_matrix <- ape::cophenetic.phylo(host_tree)

}
\references{

}
\keyword{datasets}

