% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_tree_prune.R
\docType{data}
\name{host_tree_prune}
\alias{host_tree_prune}
\title{Pylogenetic tree example on geotax.}
\format{
An object of class \code{"phylo"}
}
\source{
\url{https://phylodiversity.net/phylomatic/}
}
\usage{
host_tree_prune
}
\description{
Is a phylogenetic tree with 206 genus of potential host trees
of bark beetles. Was construct with phylomatic and prunned according
host points.
}
\keyword{datasets}
