% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Generate an incidence matrix from interaction data base}
\usage{
incidence(db)
}
\arguments{
\item{db}{A data base with two columns, interactuans and interactors}
}
\description{
Generate an incidence matrix from interaction data base
}
\examples{
incidence (db = inter_data)
}

