% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_data.R
\docType{data}
\name{inter_data}
\alias{inter_data}
\title{inter_data interaction data on geotax}
\format{An object of class \code{data.frame} with 565 rows and 2 columns.}
\source{
\href{http://barkbeetles.info/about.php}{web page}
}
\usage{
data(inter_data)
}
\description{
Data from bark beetles and their host to build an incidence matrix and a host phylogenetic tree
}
\examples{
data(inter_data)
i_matrix <- incidence(inter_data)
}
\references{
Athkinson Thomas.
#'
}
\keyword{datasets}

