% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg_boostrap.R
\name{log_reg_boostrap}
\alias{log_reg_boostrap}
\title{Bootstrap logistic regression coefficients}
\usage{
log_reg_boostrap(incidence, phydist, n)
}
\arguments{
\item{incidence}{Incidence binary matrix.}

\item{phydist}{Phylogenetic distance matrix.}

\item{n}{Number of bootstrap iterations.}
}
\value{
A named numeric vector of mean coefficients (old format).
}
\description{
Deprecated. Use \code{\link{bootstrap_geotax_model}} instead.
}
\examples{
\dontrun{
incidence_matrix <- get_incidence_matrix(beetleTreeInteractions)
incidence_matrix <- incidence_matrix[, colnames(phy_dist)]
log_reg_boostrap(incidence_matrix, phy_dist, 10)
}
}
