% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg_coef.R
\name{log_reg_coef}
\alias{log_reg_coef}
\title{Logistic regression coefficients}
\usage{
log_reg_coef(incidence, phydist)
}
\arguments{
\item{incidence}{Incidence binary matrix.}

\item{phydist}{Phylogenetic distance matrix.}
}
\value{
A data.frame of logistic regression coefficients.
}
\description{
Deprecated. Use \code{\link{fit_geotax_model}} instead.
}
\examples{
\dontrun{
incidence_matrix <- get_incidence_matrix(beetleTreeInteractions)
incidence_matrix <- incidence_matrix[, colnames(phy_dist)]
log_reg_coef(incidence_matrix, phy_dist)
}
}
