% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\docType{data}
\name{mexico}
\alias{mexico}
\title{Shapefile of Mexico.}
\format{
An object of class \code{"SpatialPolygonsDataFrame"};
}
\source{
\url{https://www.inegi.org.mx/app/mapa/espacioydatos/default.aspx}
}
\usage{
mexico
}
\description{
A shapefile of Mexico.
}
\keyword{datasets}
