% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_raster_point.R
\name{null_raster_point}
\alias{null_raster_point}
\title{Create a point matrix from a shape file with an arbitrary resolution}
\usage{
null_raster_point(shp, res)
}
\arguments{
\item{shp}{A shape file.}

\item{res}{Resolution in geographical degrees.}
}
\value{
A matrix of null raster points.
}
\description{
Deprecated. Use \code{\link{make_null_raster_points}} instead.
}
\examples{
\dontrun{
null_raster_point(shp = mexico, res = 1)
}
}
