% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_raster_polygon.R
\name{null_raster_polygon}
\alias{null_raster_polygon}
\title{Create a Spatial Polygon Data Set from a shape file}
\usage{
null_raster_polygon(shp, res)
}
\arguments{
\item{shp}{A shape file.}

\item{res}{Resolution in geographical degrees.}
}
\value{
A SpatialPolygonsDataFrame.
}
\description{
Deprecated. Use \code{\link{make_null_raster_polygons}} instead.
}
\examples{
\dontrun{
null_raster_polygon(shp = mexico, res = 1)
}
}
