% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.R
\name{PAM}
\alias{PAM}
\title{Create a Presence Absence Matrix}
\usage{
PAM(shp, points, res)
}
\arguments{
\item{shp}{A shape file.}

\item{points}{Data frame with sp long lat (3 columns).}

\item{res}{Resolution in geographical degrees.}
}
\value{
A tibble PAM.
}
\description{
Deprecated. Use \code{\link{build_presence_absence_matrix}} instead.
}
\examples{
\dontrun{
PAM(shp = mexico, points = host_points_mexico, res = 3)
}
}
