% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_logit.R
\name{prob_logit}
\alias{prob_logit}
\title{Predict probabilities from logistic regression coefficients}
\usage{
prob_logit(coef, dist)
}
\arguments{
\item{coef}{A vector of two elements (intercept and slope).}

\item{dist}{A numeric vector of distances.}
}
\value{
A vector of probabilities.
}
\description{
Deprecated. Use \code{\link{predict_geotax_probability}} instead.
}
\examples{
\dontrun{
prob_logit(c(1, -0.5), seq(0, 10, 1))
}
}
