% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{richness}
\alias{richness}
\title{Create a richness raster from species occurrens along geographical area}
\usage{
richness(shp, points, res)
}
\arguments{
\item{shp}{A shape file to delimit geographical area of the raster}

\item{points}{needs data frame with sp long lat}

\item{res}{A resolution vector of raster in geographical degrees}
}
\description{
Create a richness raster from species occurrens along geographical area
}
\examples{
richness(shp = world, points = host_points, res = 3)
}

