% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_PAM.R
\name{richness_PAM}
\alias{richness_PAM}
\title{Create a richness raster from Presence-Absence-Matrix (PAM) along geographical area}
\usage{
richness_PAM(shp, PAM, res)
}
\arguments{
\item{shp}{A shape file to delimit geographical area of the raster}

\item{PAM}{A Presence Absence Matrix}

\item{res}{A resolution vector of raster in geographical degrees}
}
\description{
Create a richness raster from Presence-Absence-Matrix (PAM) along geographical area
}
\examples{
richness_PAM(shp = world, PAM = PAM(world, host_points, 3), res = 3)
}

