% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_table.R
\docType{data}
\name{tax_table}
\alias{tax_table}
\title{tax_table taxonomical host data on geotax}
\format{An object of class \code{data.frame} with 565 rows and 5 columns.}
\source{
\href{http://barkbeetles.info/about.php}{web page}
}
\usage{
data(tax_table)
}
\description{
Data of taxonomic hierarchy from bark beetles host to build a host taxonomic tree
}
\examples{
data(tax_table)
}
\references{
Athkinson Thomas.
}
\keyword{datasets}

