% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_tree.R
\name{tax_tree}
\alias{tax_tree}
\title{Tax_tree Computes taxonomic tree from taxonomic table.
Needs a data frame witch taxonomic range as columns.}
\usage{
tax_tree(tax_table)
}
\arguments{
\item{tax_table}{A data frame with taxonomic information. Use taxonomic range as columns.}
}
\description{
Tax_tree Computes taxonomic tree from taxonomic table.
Needs a data frame witch taxonomic range as columns.
}
\examples{
tax_tree(tax_table)

}

