% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_data.R
\docType{data}
\name{world}
\alias{world}
\title{world shape file data on geotax}
\format{An object of class \code{"SpatialPolygonsDataFrame"};}
\source{

}
\usage{
data(world)
}
\description{
Spatial polygons data frame from world
}
\examples{
data(world)
r_world_null_1 <- null_raster_point(world, 1)
\donttest{plot(world)}
}
\references{

}
\keyword{datasets}

