% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incidence_matrix.R
\name{get_incidence_matrix}
\alias{get_incidence_matrix}
\title{Generate an incidence matrix from a interaction table}
\usage{
get_incidence_matrix(db, returnDataFrame = FALSE)
}
\arguments{
\item{db}{A data.frame or matrix with two columns,
for example host and parasites}

\item{returnDataFrame}{Logical, by default if FALSE and the function
returns a Matrix}
}
\value{
An incidence matrix or data.frame (tibble) with rows of one
kind (for example host) and columns with the other (for example parasites)
filled with ones and zeros.
}
\description{
This function creates an incidence matrix
}
\examples{
get_incidence_matrix(beetleTreeInteractions)
get_incidence_matrix(beetleTreeInteractions,
returnDataFrame = TRUE)
}
