% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_reg_coefficients.R
\name{get_log_reg_coefficients}
\alias{get_log_reg_coefficients}
\title{Logistic regression coeficients}
\usage{
get_log_reg_coefficients(incidence, phydist)
}
\arguments{
\item{incidence}{Incidence binary matrix between interactions (example
host - parasite system)}

\item{phydist}{Phylogenetic distance matrix among host}
}
\value{
A table with logistic regression coefficients
}
\description{
Calculates the coefficients of a logistic
regression. Use an incidence matrix of pest - host interaction and a
phylogenetical distance matrix of host.
}
\examples{
phy_dist <- ape::cophenetic.phylo(host_tree)
incidence_matrix <-  get_incidence_matrix(beetleTreeInteractions)
incidence_matrix <- incidence_matrix[ ,colnames(phy_dist)]
get_log_reg_coefficients(incidence_matrix, phy_dist)
}
