% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomical_tree.R
\name{get_taxonomical_tree}
\alias{get_taxonomical_tree}
\title{Create a taxonomic tree (cladogram) given a taxonomic table}
\usage{
get_taxonomical_tree(tax_table, power = 1)
}
\arguments{
\item{tax_table}{A data frame (or tibble) with taxonomic information.}

\item{power}{The power at which heights must be raised according to
Grafen's (1989) computation of branch lengths.
Use taxonomic hierarchy (class, order, family genus, species) as columns.
Default is 1 to get distances as integer numbers.}
}
\value{
A cladogram (tree) from taxonomic table
}
\description{
Internally use the as.phylo.formula from the ape R package. Internally parse
the name of the taxonomical hierarchy to a formula. Works with a table
which columns are the taxonomic hierarchy from left to right.
}
\details{
Use the same algorithm in compute.brlen from ape package
}
\examples{
get_taxonomical_tree(tax_table)
}
\references{
\url{https://search.r-project.org/CRAN/refmans/ape/html/compute.brlen.html}
}
