% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_points.R
\docType{data}
\name{host_points}
\alias{host_points}
\title{host_points host ocurrence data on geotax}
\format{An object of class \code{"data.frame"}}
\usage{
data(host_points)
}
\description{
Data of host presence points
}
\examples{
data(host_points)
PAM_host_world <- PAM(world, host_points, 3)

}
\references{

}
\keyword{datasets}

