% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_tree.R
\docType{data}
\name{host_tree}
\alias{host_tree}
\title{Pylogenetic tree example on geotax.}
\format{
An object of class \code{"phylo"}
}
\source{
\url{https://phylodiversity.net/phylomatic/}
}
\usage{
host_tree
}
\description{
Is a phylogenetic tree with 235 genus of potential host trees
of bark beetles. Was construct with phylomatic. Was generated with
phylomatic
}
\keyword{datasets}
