% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg_boostrap.R
\name{log_reg_boostrap}
\alias{log_reg_boostrap}
\title{Logistic regression coeficients.
Calculates the coefficients of a logistic regression.
As input use an incidence matrix of host - parasite interaction
and a phylogenetical distance matrix of host.
It replicates the calculation of the coefficients 1000 times
and gets the central tendency of the distribution of each coefficient.}
\usage{
log_reg_boostrap(incidence, phydist, n)
}
\arguments{
\item{incidence}{Incidence binary matrix between pest - host interaction}

\item{phydist}{Phylogenetic host distance matrix}

\item{n}{Number of boostrap iteration}
}
\description{
Logistic regression coeficients.
Calculates the coefficients of a logistic regression.
As input use an incidence matrix of host - parasite interaction
and a phylogenetical distance matrix of host.
It replicates the calculation of the coefficients 1000 times
and gets the central tendency of the distribution of each coefficient.
}
\examples{
phy_dist <- ape::cophenetic.phylo(host_tree)
incidence_matrix <-  get_incidence_matrix(beetleTreeInteractions)
incidence_matrix <- incidence_matrix[ , colnames(phy_dist)]
log_reg_boostrap(incidence_matrix, phy_dist, 10)
}
