% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_null_raster_points.R
\name{make_null_raster_points}
\alias{make_null_raster_points}
\title{Create null raster points from a shape file}
\usage{
make_null_raster_points(shp, res)
}
\arguments{
\item{shp}{A spatial object (shape file).}

\item{res}{Positive numeric. Raster resolution in geographical degrees.}
}
\value{
A matrix of null raster points.
}
\description{
Creates a point matrix from a shape file at an arbitrary resolution.
This is the v2 replacement for \code{null_raster_point()}.
}
\examples{
make_null_raster_points(shp = mexico, res = 1)
}
\seealso{
\code{\link{null_raster_point}} (deprecated)
}
