% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_null_raster_polygons.R
\name{make_null_raster_polygons}
\alias{make_null_raster_polygons}
\title{Create null raster polygons from a shape file}
\usage{
make_null_raster_polygons(shp, res)
}
\arguments{
\item{shp}{A spatial object (shape file).}

\item{res}{Positive numeric. Raster resolution in geographical degrees.}
}
\value{
A SpatialPolygonsDataFrame of null raster polygons.
}
\description{
Creates a spatial polygon data set from a shape file at an arbitrary
resolution. This is the v2 replacement for \code{null_raster_polygon()}.
}
\examples{
make_null_raster_polygons(shp = mexico, res = 1)
}
\seealso{
\code{\link{null_raster_polygon}} (deprecated)
}
