% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_raster_polygon.R
\name{null_raster_polygon}
\alias{null_raster_polygon}
\title{Create a Spatial polygon Data Set from a shape file
with an arbitrary resolution}
\usage{
null_raster_polygon(shp, res)
}
\arguments{
\item{shp}{A shape file}

\item{res}{A resolution vector of the null raster in geographical degrees}
}
\value{
a null polygon
}
\description{
Create a Spatial polygon Data Set from a shape file
with an arbitrary resolution
}
\examples{
null_raster_polygon(shp = mexico, res = 1)
}
