% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.R
\name{PAM}
\alias{PAM}
\title{Create a Presence Absence Matrix}
\usage{
PAM(shp, points, res)
}
\arguments{
\item{shp}{A shape file}

\item{points}{Data frame with sp long lat}

\item{res}{A resolution vector of the null raster in geographical degrees}
}
\description{
Create a Presence Absence Matrix
}
\examples{
PAM(shp = world, points = host_points, res = 3)
}

