% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_dist.R
\docType{data}
\name{phy_dist}
\alias{phy_dist}
\title{phy_dist phylogenetical distance matrix on geotax}
\format{An object of class \code{matrix} with 235 rows and 235 columns.}
\usage{
data(phy_dist)
}
\description{
Matrix with phylogenetical distance between tips of taxonomic tree
}
\examples{
data(phy_dist)

}
\keyword{datasets}

