% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_taxonomic_tree.R
\name{prepare_taxonomic_tree}
\alias{prepare_taxonomic_tree}
\title{Prepare a taxonomic tree (cladogram) from a taxonomic table}
\usage{
prepare_taxonomic_tree(tax_table, power = 1)
}
\arguments{
\item{tax_table}{A data.frame with taxonomic information. Columns must include
at least 3 of: class, order, family, genus, species.}

\item{power}{Numeric. The power at which heights must be raised according to
Grafen's (1989) computation of branch lengths. Default is 1.}
}
\value{
A phylo object (cladogram) derived from the taxonomic table.
}
\description{
Creates a phylogenetic tree from a taxonomic table using taxonomic ranks
as a hierarchy. This is the v2 replacement for \code{get_taxonomical_tree()}.
}
\details{
Uses the same algorithm as \code{\link[ape]{compute.brlen}} from the
ape package.
}
\examples{
prepare_taxonomic_tree(tax_table)
}
\references{
\url{https://search.r-project.org/CRAN/refmans/ape/html/compute.brlen.html}
}
\seealso{
\code{\link{get_taxonomical_tree}} (deprecated)
}
