% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_logit.R
\name{prob_logit}
\alias{prob_logit}
\title{prob_logit}
\usage{
prob_logit(coef, dist)
}
\arguments{
\item{coef}{Coefficients of logistic regresion}

\item{dist}{A distance matrix to perform regression}
}
\description{
prob_logit
}
\examples{
prob_logit(coef, phy_dist)
}

