% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_table.R
\docType{data}
\name{tax_table}
\alias{tax_table}
\title{Taxonomical host data on geotax}
\format{
A data frame with 3254 rows and 3 variables:
\describe{
  \item{class}{class of tree host of bark beetle}
  \item{order}{order of tree host of bark beetle}
  \item{family}{family of tree host of bark beetle}
  \item{genus}{genus of tree host of bark beetle}
  \item{species}{species of tree host of bark beetle}
}
}
\source{
\url{https://www.barkbeetles.info/}
}
\usage{
tax_table
}
\description{
Data of taxonomic hierarchy from bark beetles host to build
a taxonomic cladogram from presence points of bark beetles.
}
\keyword{datasets}
